feat(wire): widen shared-data operands to 24 bits (BYTECODE_VERSION 1 -> 2)
Raise the shared-data segment ceiling from 64 KB to 16 MB by widening the
byte-offset, unified data-slot index, and indexed-array length operands of
GetData/SetData/GetDataIndexed/SetDataIndexed from sixteen to twenty-four bits.
This unblocks compiling programs whose shared segment exceeds 64 KB (the
self-hosted parser stage, whose 67 KB source and ~80 KB packed token stream both
overflow the old u16 fields, is the motivating case, but the change benefits all
programs).

The wire already affords the width with no record growth: GetData/SetData use
the three inline operand bytes (previously two, byte three padding) as a 24-bit
little-endian value, and GetDataIndexed/SetDataIndexed pack two 24-bit values
into the operand-pool entry's six-byte payload under a new (u24, u24) tag 0x04
(tag 0x01 stays 16-bit for the nested-composite and IsEnum shapes). The four-byte
opcode record and eight-byte pool entry are unchanged, so bytecode does not grow
-- the rad-hard minimal-ISA and the 32-bit embedded FLASH budget are preserved.

In memory the four ops carry u32 operands (validated < 2^24 at compile and at
load via MAX_DATA_ADDR), SharedSlotLayout.offset becomes u32, and the VM's shared
and private slot counts widen to u32 so a >64K-slot module is counted without
saturation. The per-level BoundsCheck operand stays u16, so a multi-dimensional
per-level dimension above 64 K is still rejected; only the outermost indexed
length uses the wider operand. WCMU is unaffected (it accounts arena and private
memory, not the host-provided shared buffer).

This is a deliberate BYTECODE_VERSION bump: V1 bytecode is rejected at load on
the version check. The golden wire fixture is regenerated (only the version byte
and CRC trailer move for a data-segment-free program). New tests: a wire
encode/decode round-trip for 24-bit values, and tests/wide_shared_segment.rs
compiling, wire-round-tripping, and running a program with a 70000-byte shared
array (byte offset and slot index past 65535) plus set_shared/get_shared at a
high slot. Docs: WIRE_FORMAT.md (Version 2, the (u24, u24) tag, the u32 offset)
and CHANGELOG.md. Incidental: cleared one pre-existing redundant-doc-link lint in
verify_typed.rs that reddened the cargo doc gate.

Gate green: default, +signatures, --all-features test suites; clippy
--tests --all-features -D warnings; fmt; cargo doc -D warnings; narrow-word-16
lib build; the detached compiler workspace builds.

+++ tests/typed_conformance.rs
@@ -118,7 +118,7 @@ fn b6_shared_slot_offset_overrun_rejected() {
         "expected a shared-slot layout entry"
     );
     // Push the first slot's offset past the end of the shared-data buffer.
-    layout.shared_layout[0].offset = (buffer + 8) as u16;
+    layout.shared_layout[0].offset = buffer + 8;
     assert!(
         rejected_by_typed_pass(&m),
         "a shared-slot offset past the buffer must be rejected by the typed pass"
